requirements.txt
torch
torchvision
transformers
kagglehub
gradio
huggingface_hub
scikit-learn
matplotlib
seaborn
tqdm
Pillow
